fix(macos): fourth wave of macOS compile fixes
- Add POSIX stubs to compat.h: SetCurrentDirectory, WIN32_FIND_DATA,
  FindFirstFile/FindNextFile/FindClose, locale APIs (GetDateFormatW,
  GetTimeFormat), EXCEPTION_POINTERS/CONTEXT, DirectX 8 type stubs
  (LPDIRECT3DDEVICE8, LPDIRECT3DVERTEXBUFFER8, etc.)
- Wrap StackDump with #ifdef _WIN32; add macOS backtrace() fallback
- Fix pointer-to-int truncation casts across 30+ files with (intptr_t)/
  (uintptr_t) intermediates
- Wrap x87 FPU control (_fpreset, _statusfp, _MCW_RC) with #ifdef _WIN32
- Wrap entire W3DVolumetricShadow.cpp and W3DProjectedShadow.cpp with
  #ifdef _WIN32 (files are all DirectX 8)
- Wrap <windows.h>, <io.h>, <mmsystem.h> includes with #ifdef _WIN32
  in W3DDisplay, W3DFileSystem, W3DMainMenu, WinMain.h
- Guard ApplicationHWnd usage sites in W3DDisplay with #ifdef _WIN32
- Fix W3DWebBrowser: Windows impl under #ifdef _WIN32, stub on macOS
- Fix W3DScene: guard direct IDirect3DDevice8 method calls with #ifdef
- Fix W3DAssetManager pointer arithmetic in lstrcpyn call
- Rename pause→s_paused in InGamePopupMessage to avoid POSIX conflict

+++ Core/GameEngine/Source/Common/System/LocalFile.cpp
@@ -454,7 +454,7 @@ Int LocalFile::writeFormat( const WideChar* format, ... )
 Int LocalFile::writeChar( const Char* character )
 {
 	if ( write( character, sizeof(Char) ) == sizeof(Char) ) {
-		return (Int)character;
+		return (Int)(uintptr_t)character;
 	}
 
 	return EOF;
@@ -467,7 +467,7 @@ Int LocalFile::writeChar( const Char* character )
 Int LocalFile::writeChar( const WideChar* character )
 {
 	if ( write( character, sizeof(WideChar) ) == sizeof(WideChar) ) {
-		return (Int)character;
+		return (Int)(uintptr_t)character;
 	}
 
 	return WEOF;
